messaging_app/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
mysqlclient==2.2.0
python-decouple==3.8
gunicorn==21.2.0
